requirements.txt
Flask==3.1.0
langchain==0.3.17
langchain_community==0.3.16
langchain_core==0.3.33
langchain_google_genai==2.0.9
langgraph==0.2.69
pandas==2.2.3
pydantic==2.10.6
python-dotenv==1.0.1
pytz==2024.2
pytz==2024.1
pytz==2025.1
Requests==2.32.3
typing_extensions==4.12.2
Werkzeug==3.1.3